方向环串行加入

+++ dh/Inc/control.h
@@ -5,9 +5,9 @@
 extern void megnet();
 extern void direction_control();
 extern int rightspeed;
-extern int rightspeedset;
+extern float rightspeedset;
 extern int leftspeed;
-extern int leftspeedset;
+extern float leftspeedset;
 extern void speed_control();
 #endif /*__ control_H */
